requirements.txt
requests
pyarrow
lxml
google-cloud-storage
